+++ ext/bcmath/libbcmath/src/bcmath.h
@@ -32,19 +32,21 @@
 #ifndef _BCMATH_H_
 #define _BCMATH_H_
 
+#include <stddef.h>
+
 typedef enum {PLUS, MINUS} sign;
 
 typedef struct bc_struct *bc_num;
 
 typedef struct bc_struct {
-	sign  n_sign;
-	int   n_len;   /* The number of digits before the decimal point. */
-	int   n_scale; /* The number of digits after the decimal point. */
-	int   n_refs;  /* The number of pointers to this number. */
-	char *n_ptr;   /* The pointer to the actual storage.
+	sign   n_sign;
+	size_t n_len;   /* The number of digits before the decimal point. */
+	size_t n_scale; /* The number of digits after the decimal point. */
+	int    n_refs;  /* The number of pointers to this number. */
+	char  *n_ptr;   /* The pointer to the actual storage.
 	                  If NULL, n_value points to the inside of another number
 	                  (bc_multiply...) and should not be "freed." */
-	char *n_value;  /* The number. Not zero char terminated.
+	char  *n_value; /* The number. Not zero char terminated.
 	                   May not point to the same place as n_ptr as
 	                   in the case of leading zeros generated. */
 } bc_struct;
@@ -84,7 +86,7 @@ typedef struct bc_struct {
 
 void bc_init_numbers(void);
 
-bc_num _bc_new_num_ex(int length, int scale, bool persistent);
+bc_num _bc_new_num_ex(size_t length, size_t scale, bool persistent);
 
 void _bc_free_num_ex(bc_num *num, bool persistent);
 
@@ -92,9 +94,9 @@ bc_num bc_copy_num(bc_num num);
 
 void bc_init_num(bc_num *num);
 
-bool bc_str2num(bc_num *num, char *str, int scale);
+bool bc_str2num(bc_num *num, char *str, size_t scale);
 
-zend_string *bc_num2str_ex(bc_num num, int scale);
+zend_string *bc_num2str_ex(bc_num num, size_t scale);
 
 void bc_int2num(bc_num *num, int val);
 
@@ -104,29 +106,29 @@ int bc_compare(bc_num n1, bc_num n2);
 
 bool bc_is_zero(bc_num num);
 
-bool bc_is_zero_for_scale(bc_num num, int scale);
+bool bc_is_zero_for_scale(bc_num num, size_t scale);
 
-bool bc_is_near_zero(bc_num num, int scale);
+bool bc_is_near_zero(bc_num num, size_t scale);
 
 bool bc_is_neg(bc_num num);
 
-void bc_add(bc_num n1, bc_num n2, bc_num *result, int scale_min);
+void bc_add(bc_num n1, bc_num n2, bc_num *result, size_t scale_min);
 
-void bc_sub(bc_num n1, bc_num n2, bc_num *result, int scale_min);
+void bc_sub(bc_num n1, bc_num n2, bc_num *result, size_t scale_min);
 
-void bc_multiply(bc_num n1, bc_num n2, bc_num *prod, int scale);
+void bc_multiply(bc_num n1, bc_num n2, bc_num *prod, size_t scale);
 
 int bc_divide(bc_num n1, bc_num n2, bc_num *quot, int scale);
 
-int bc_modulo(bc_num num1, bc_num num2, bc_num *resul, int scale);
+int bc_modulo(bc_num num1, bc_num num2, bc_num *resul, size_t scale);
 
-int bc_divmod(bc_num num1, bc_num num2, bc_num *quo, bc_num *rem, int scale);
+int bc_divmod(bc_num num1, bc_num num2, bc_num *quo, bc_num *rem, size_t scale);
 
-zend_result bc_raisemod(bc_num base, bc_num expo, bc_num mo, bc_num *result, int scale);
+zend_result bc_raisemod(bc_num base, bc_num expo, bc_num mo, bc_num *result, size_t scale);
 
-void bc_raise(bc_num num1, bc_num num2, bc_num *resul, int scale);
+void bc_raise(bc_num num1, bc_num num2, bc_num *resul, size_t scale);
 
-bool bc_sqrt(bc_num *num, int scale);
+bool bc_sqrt(bc_num *num, size_t scale);
 
 void bc_out_num(bc_num num, int o_base, void (* out_char)(char), bool leading_zero);
