+++ ext/bcmath/libbcmath/src/divmod.c
@@ -30,6 +30,7 @@
 *************************************************************************/
 
 #include "bcmath.h"
+#include <stddef.h>
 
 
 /* Division *and* modulo for numbers.  This computes both NUM1 / NUM2 and
@@ -37,11 +38,11 @@
    is NULL then that store will be omitted.
  */
 
-int bc_divmod(bc_num num1, bc_num num2, bc_num *quot, bc_num *rem, int scale)
+int bc_divmod(bc_num num1, bc_num num2, bc_num *quot, bc_num *rem, size_t scale)
 {
 	bc_num quotient = NULL;
 	bc_num temp;
-	int rscale;
+	size_t rscale;
 
 	/* Check for correct numbers. */
 	if (bc_is_zero (num2)) {
@@ -73,7 +74,7 @@ int bc_divmod(bc_num num1, bc_num num2, bc_num *quot, bc_num *rem, int scale)
 /* Modulo for numbers.  This computes NUM1 % NUM2  and puts the
    result in RESULT.   */
 
-int bc_modulo(bc_num num1, bc_num num2, bc_num *result, int scale)
+int bc_modulo(bc_num num1, bc_num num2, bc_num *result, size_t scale)
 {
 	return bc_divmod(num1, num2, NULL, result, scale);
 }
