+++ ext/bcmath/libbcmath/src/zero.c
@@ -30,13 +30,14 @@
 *************************************************************************/
 
 #include "bcmath.h"
+#include <stddef.h>
 #include <stdbool.h>
 
 /* In some places we need to check if the number NUM is zero. */
 
-bool bc_is_zero_for_scale (bc_num num, int scale)
+bool bc_is_zero_for_scale (bc_num num, size_t scale)
 {
-	int  count;
+	size_t count;
 	char *nptr;
 
 	/* Quick check. */
